fix smoke test

+++ sky/jobs/controller.py
@@ -603,7 +603,12 @@ def _cleanup(job_id: int, dag_yaml: str):
         # mounts.
         for file_mount in (task.file_mounts or {}).values():
             try:
-                if not data_utils.is_cloud_store_url(file_mount):
+                # For consolidation mode, there is no two-hop file mounts
+                # and the file path here represents the real user data.
+                # We skip the cleanup for consolidation mode.
+                if not data_utils.is_cloud_store_url(
+                        file_mount
+                ) and not managed_job_utils.is_consolidation_mode():
                     path = os.path.expanduser(file_mount)
                     if os.path.isdir(path):
                         shutil.rmtree(path)
